+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add Rknn1, Rknn2 and Rknn4 as models.
 * Remove neuralode_type1 and keep only neuralode_type2 as neuralODE from AutomationLabsIdentification.jl.
 * Remove Mosektools as deps and mosek_solver_def type.
+* Remove HiGHS as deps and highs_solver_def type.
 * Remove windows os ci.
 * Tests improvement.